+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Fix lalrpop builds when extra folders exist.
 * Refactor operator metrics collection to eliminate clones in most cases.
 * Fix systemd spec file to load tremor files in `/etc/tremor/config` [#784](https://github.com/tremor-rs/tremor-runtime/issues/784)
+* Do not crash when we can not execute a config file [#792](https://github.com/tremor-rs/tremor-runtime/issues/792)
 
 ## 0.10.2